servers: benchmark: properly computing stripes per request now due to floating point bug

+++ java/servers/src/org/xtreemfs/common/benchmark/UnalignedSequentialWriteBenchmark.java
@@ -43,7 +43,7 @@ public class UnalignedSequentialWriteBenchmark extends SequentialBenchmark {
         long byteCounter = 0;
 
         for (long j = 0; !cancelled && j < numberOfBlocks; j++) {
-            long stripesPerRequest = (long) Math.ceil(requestSize / config.getStripeSizeInBytes());
+            long stripesPerRequest = (long) Math.ceil((double) requestSize / (double) config.getStripeSizeInBytes());
             long nextOffset = j * stripesPerRequest * config.getStripeSizeInBytes();
             random.nextBytes(data);
             byteCounter += fileHandle.write(config.getUserCredentials(), data, requestSize, nextOffset);
